Adjust array_cwise for 32-bit arm.
The `atan` test is currently failing because of 32-bit arm's FTZ behavior.
The polynomial expansion for `atan` leads to a flushed zero result even
when the true result is >= `min` for certain small inputs, since each
term in the polynomial sum is < `min`.  Slightly modified the special
value inputs to avoid this case.